# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collection/value-history:
    get:
      operationId: getCollectionValueHistory
      tags: [portfolio]
      summary: Get portfolio value history derived from collection holdings
      parameters:
        - { name: range, in: query, schema: { type: string, enum: [1D, 7D, 1M, 3M, 6M, 1Y, ALL], default: ALL } }
        - { name: displayCurrency, in: query, schema: { type: string, default: AUD } }
      responses:
        "200":
          description: Historical portfolio points from persisted holdings and prices.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionValueHistory" }
        "401": { description: Authentication required. }
...
components:
  schemas:
...
    PortfolioValuePoint:
      type: object
      required: [date, value, currency]
      properties:
        date: { type: string, format: date }
        value: { type: number }
        currency: { type: string }
    CollectionValueHistory:
      type: object
      required: [range, currency, points, chartData, historyAvailable]
      properties:
        range:
          type: string
          enum: [1D, 7D, 1M, 3M, 6M, 1Y, ALL]
        currency: { type: string }
        points:
          type: array
          items: { $ref: "#/components/schemas/PortfolioValuePoint" }
        chartData:
          type: object
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/PortfolioValuePoint" }
        historyAvailable: { type: boolean }
        historyUnavailableReason: { type: [string, "null"] }
...